Support custom build_file_content
This adds a string parameter `build_file_content` to the `deb_package`
rule to allow for customization. One use case are packages that have
files under `usr/include` or `usr/lib` that are not allowed in `hdrs` or
`srcs`, for example `*.cmake`. With a custom `build_file_content`, we
can fix that.